+++ green/src/main/java/com/blockstream/green/ui/AppViewModel.kt
@@ -1,27 +1,20 @@
 package com.blockstream.green.ui
 
-import androidx.arch.core.util.Function
-import androidx.lifecycle.Lifecycle
-import androidx.lifecycle.LifecycleOwner
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
+import androidx.lifecycle.*
 import com.blockstream.DeviceBrand
 import com.blockstream.gdk.data.Device
+import com.blockstream.green.data.AppEvent
+import com.blockstream.green.ui.devices.DeviceInfoViewModel
 import com.blockstream.green.utils.ConsumableEvent
 import com.greenaddress.greenapi.HWWallet
 import com.greenaddress.greenapi.HWWalletBridge
+import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
 import io.reactivex.rxjava3.core.Single
-
+import io.reactivex.rxjava3.core.SingleEmitter
 import io.reactivex.rxjava3.disposables.CompositeDisposable
-import androidx.lifecycle.LifecycleRegistry
-import com.blockstream.green.data.AppEvent
 
 
 open class AppViewModel : ViewModel(), HWWalletBridge, LifecycleOwner {
-
-//    enum class Event {
-//        NAVIGATE, RENAME_WALLET, DELETE_WALLET, RENAME_ACCOUNT, ACK_MESSAGE
-//    }
 
     internal val disposables = CompositeDisposable()
     private var lifecycleRegistry: LifecycleRegistry? = null
@@ -32,21 +25,29 @@ open class AppViewModel : ViewModel(), HWWalletBridge, LifecycleOwner {
 
     val onDeviceInteractionEvent = MutableLiveData<ConsumableEvent<Device>>()
 
+    var requestPinMatrixEmitter: SingleEmitter<String>? = null
+    var requestPinPassphraseEmitter: SingleEmitter<String>? = null
+
     override fun interactionRequest(hw: HWWallet?) {
         hw?.let {
             onDeviceInteractionEvent.postValue(ConsumableEvent(it.device))
         }
     }
 
-    // The following two methods are not needed
-    // it will be remove in the next iteration on simplifying
-    // hardware wallet interfaces
     override fun requestPinMatrix(deviceBrand: DeviceBrand?): Single<String> {
-        TODO("Not yet implemented")
+        onEvent.postValue(ConsumableEvent(AppFragment.DeviceRequestEvent.RequestPinMatrix))
+
+        return Single.create<String> { emitter ->
+            requestPinMatrixEmitter = emitter
+        }.subscribeOn(AndroidSchedulers.mainThread())
     }
 
     override fun requestPassphrase(deviceBrand: DeviceBrand?): Single<String> {
-        TODO("Not yet implemented")
+        onEvent.postValue(ConsumableEvent(AppFragment.DeviceRequestEvent.RequestPassphrase))
+
+        return Single.create<String> { emitter ->
+            requestPinPassphraseEmitter = emitter
+        }.subscribeOn(AndroidSchedulers.mainThread())
     }
 
     override fun onCleared() {