+++ green/src/main/java/com/blockstream/green/ui/WalletFragment.kt
@@ -121,6 +121,8 @@ abstract class WalletFragment<T : ViewDataBinding> constructor(
 
     abstract fun getWalletViewModel(): AbstractWalletViewModel
 
+    override fun getAppViewModel() = getWalletViewModel()
+
     override fun onResume() {
         super.onResume()