+++ CHANGELOG.md
@@ -22,6 +22,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fix 2FA popup truncated on small screen
 - Fix BIP21 uris 
 - Fix crash when deleting a wallet
+- Fix Trezor One with Passphrase enabled
 
 ## [3.7.3] - 2021-09-30
 ### Changed